feat(atoms): rename Switch to Checkbox with switch/checkbox appearances
Adds @radix-ui/react-checkbox and a CheckIcon. The `switch` appearance is
the former Switch verbatim (the default, so existing toggles are visually
unchanged); the new `checkbox` appearance (role="checkbox") unblocks the
registration mailing-list consent field. Migrates the sole consumer,
SearchBar, to <Checkbox appearance="switch"> — pixel-identical.

Part of #25
