+++ package.json
@@ -25,6 +25,7 @@
     "@mapbox/search-js-react": "1.0.0-beta.24",
     "@r2wc/react-to-web-component": "^2.0.4",
     "@radix-ui/colors": "^3.0.0",
+    "@radix-ui/react-checkbox": "^1.3.6",
     "@radix-ui/react-dialog": "^1.1.18",
     "@radix-ui/react-label": "^2.1.11",
     "@radix-ui/react-select": "^2.3.2",
