+++ src/components/atoms/Checkbox/index.ts
@@ -1,0 +1,2 @@
+export { Checkbox } from './Checkbox'
+export type { CheckboxProps } from './Checkbox'
