--- src/components/atoms/Switch/index.ts
@@ -1,2 +1,0 @@
-export { Switch } from './Switch'
-export type { SwitchProps } from './Switch'
